+++ extensions/threads/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Threads Changelog
 
+## [Feature] - {PR_MERGE_DATE}
+
+- Added the `Download Threads Video` command to download videos from Threads.
+
 ## [Initial Version] - 2024-08-21